Add core, core-min, and plugins/* exports for better bundling internally
