+++ src/Exports/BundlePlugin.js
@@ -1,0 +1,1 @@
+export { default as BundlePlugin } from "@11ty/eleventy-plugin-bundle";
