+++ src/Exports/HtmlBasePlugin.js
@@ -1,0 +1,1 @@
+export { default as HtmlBasePlugin } from "../Plugins/HtmlBasePlugin.js";
