+++ src/Exports/I18nPlugin.js
@@ -1,0 +1,1 @@
+export { default as I18nPlugin } from "../Plugins/I18nPlugin.js";
