+++ src/Exports/IdAttributePlugin.js
@@ -1,0 +1,1 @@
+export { IdAttributePlugin } from "../Plugins/IdAttributePlugin.js";
