+++ src/Exports/InputPathToUrlPlugin.js
@@ -1,0 +1,1 @@
+export { TransformPlugin as InputPathToUrlTransformPlugin } from "../Plugins/InputPathToUrl.js";
