+++ src/Exports/RenderPlugin.js
@@ -1,0 +1,1 @@
+export { default as RenderPlugin } from "../Plugins/RenderPlugin.js";
